+++ packages/core/chat/mutations.ts
@@ -1,6 +1,6 @@
 import { useMutation, useQueryClient } from "@tanstack/react-query";
-import { api } from "@multica/core/api";
-import { useWorkspaceId } from "@multica/core/hooks";
+import { api } from "../api";
+import { useWorkspaceId } from "../hooks";
 import { chatKeys } from "./queries";
 
 export function useCreateChatSession() {
