+++ packages/core/package.json
@@ -34,6 +34,10 @@
     "./inbox/queries": "./inbox/queries.ts",
     "./inbox/mutations": "./inbox/mutations.ts",
     "./inbox/ws-updaters": "./inbox/ws-updaters.ts",
+    "./chat": "./chat/index.ts",
+    "./chat/queries": "./chat/queries.ts",
+    "./chat/mutations": "./chat/mutations.ts",
+    "./chat/store": "./chat/store.ts",
     "./runtimes": "./runtimes/index.ts",
     "./runtimes/queries": "./runtimes/queries.ts",
     "./runtimes/mutations": "./runtimes/mutations.ts",
