+++ packages/views/chat/components/chat-input.tsx
