+++ packages/views/chat/index.ts
@@ -1,3 +1,2 @@
 export { ChatFab } from "./components/chat-fab";
 export { ChatWindow } from "./components/chat-window";
-export { useChatStore } from "./store";
