+++ packages/views/package.json
@@ -31,6 +31,7 @@
     "./layout": "./layout/index.ts",
     "./auth": "./auth/index.ts",
     "./search": "./search/index.ts",
+    "./chat": "./chat/index.ts",
     "./settings": "./settings/index.ts"
   },
   "dependencies": {
